This time for real :(

+++ InfServer/Protocol/Packets/Game/Update/SC_PlayerTeamUpdate.cs
@@ -76,7 +76,7 @@ namespace InfServer.Protocol
 					Write(vehicle._state.positionZ);
 					Write(vehicle._state.yaw);
 					Write((UInt16)vehicle._state.direction);
-					Write(vehicle._state.unk1);
+                    Write((byte)vehicle._state.unk1);
 					break;
 
 				case VehInfo.Types.Dependent:
@@ -90,7 +90,7 @@ namespace InfServer.Protocol
 					Write(vehicle._state.pitch);
 					Write(vehicle._state.yaw);
 					Write(vehicle._state.yaw);
-					Write(vehicle._state.unk1);
+                    Write((byte)vehicle._state.unk1);
 					break;
 
 				case VehInfo.Types.Computer:
@@ -104,7 +104,7 @@ namespace InfServer.Protocol
 					Write((short)0);
 					Write(vehicle._state.fireAngle);	//Base yaw?
 					Write(vehicle._state.fireAngle);	//Turret yaw?
-					Write((short)206);
+					Write((ushort)206);
 					break;
 			}
